make an api stub for non-backend mode

+++ src/actions/products.js
@@ -1,5 +1,7 @@
 import axios from 'axios';
 import { toast } from 'react-toastify';
+import mock from "../pages/products/backendMock";
+import config from "../config";
 
 export const RECEIVED_PRODUCTS = 'RECEIVED_PRODUCTS';
 export const RECEIVING_PRODUCTS = 'RECEIVING_PRODUCTS';
@@ -13,24 +15,41 @@ export const RECEIVED_IMAGES = 'RECEIVED_IMAGES';
 
 export function getProductsRequest() {
     return (dispatch) => {
-        dispatch(receivingProducts());
-        axios.get('/products').then(res => {
-            dispatch(receiveProducts(res.data));
-        })
+        // We check if app runs with backend mode
+        if (!config.isBackend) {
+            dispatch(receiveProducts(mock));
+        }
+
+        else {
+            dispatch(receivingProducts());
+            axios.get('/products').then(res => {
+                dispatch(receiveProducts(res.data));
+            })
+        }
     };
 }
 
 export function loadProductRequest(id) {
     return (dispatch) => {
-        dispatch(receivingProduct());
-        axios.get('/products/' + id).then(res => {
-            dispatch(receiveProduct(res.data));
-        })
+        // We check if app runs with backend mode
+        if (!config.isBackend) {
+            dispatch(receiveProduct(mock.find(arr => arr.id === id)));
+        }
+
+        else {
+            dispatch(receivingProduct());
+            axios.get('/products/' + id).then(res => {
+                dispatch(receiveProduct(res.data));
+            })
+        }
     };
 }
 
 export function updateProductRequest(product) {
     return (dispatch) => {
+        // We check if app runs with backend mode
+        if (!config.isBackend) return;
+
         dispatch(updatingProduct());
         axios.put('/products/' + product.id, product).then(res => {
             dispatch(updateProduct(res.data));
@@ -41,6 +60,9 @@ export function updateProductRequest(product) {
 
 export function createProductRequest(payload) {
     return (dispatch) => {
+        // We check if app runs with backend mode
+        if (!config.isBackend) return;
+
         dispatch(updatingProduct());
         axios.post('/products', payload.product).then(res => {
             dispatch(updateProduct(res.data));
@@ -52,6 +74,9 @@ export function createProductRequest(payload) {
 
 export function deleteProductRequest(payload) {
     return (dispatch) => {
+        // We check if app runs with backend mode
+        if (!config.isBackend) return;
+
         dispatch(deletingProduct(payload));
         axios.delete('/products/' + payload.id).then(res => {
             dispatch(deleteProduct({id: payload.id}));
@@ -65,6 +90,9 @@ export function deleteProductRequest(payload) {
 
 export function getProductsImagesRequest(payload) {
   return (dispatch) => {
+    // We check if app runs with backend mode
+    if (!config.isBackend) return;
+
     axios.get('/products/images-list').then(res => {
       dispatch(receiveProductImages(res.data));
       if (!payload.img && res.data.length) {
